+++ src/Parser/PublicationParser.php
@@ -15,7 +15,7 @@ use Symfony\Component\DomCrawler\Crawler;
 class PublicationParser extends BaseParser implements Parser
 {
 
-    const GSCHOLAR_XPATH_PUBLICATIONS = '//table[@id="gsc_a_t"]/tbody[@id="gsc_a_b"]/tr[@class="gsc_a_tr"]';
+    const GSCHOLAR_XPATH = '//table[@id="gsc_a_t"]/tbody[@id="gsc_a_b"]/tr[@class="gsc_a_tr"]';
 
     const GSCHOLAR_CSS_CLASS_REFERENCE = 'gsc_a_t';
     const GSCHOLAR_CSS_CLASS_REFERENCE_TITLE = 'gsc_a_at';
@@ -31,7 +31,7 @@ class PublicationParser extends BaseParser implements Parser
         $data = [];
 
         /** @var Crawler $crawlerPublications */
-        $crawlerPublications = $this->crawler->filterXPath(self::GSCHOLAR_XPATH_PUBLICATIONS);
+        $crawlerPublications = $this->crawler->filterXPath(self::GSCHOLAR_XPATH);
 
         /** @var DOMElement $domPublication */
         foreach ($crawlerPublications as $domPublication) {
@@ -71,7 +71,7 @@ class PublicationParser extends BaseParser implements Parser
      */
     private function parsePublicationTitle(DOMElement $node)
     {
-        $title = $gScholarPath = $authors = $publisherDetails = '';
+        $title = $publicationPath = $authors = $publisherDetails = '';
         $childNodeIndex = 0;
 
         foreach ($node->childNodes as $childNode) {
@@ -79,7 +79,7 @@ class PublicationParser extends BaseParser implements Parser
 
             if ($cssClass === self::GSCHOLAR_CSS_CLASS_REFERENCE_TITLE) {
                 $title = $childNode->textContent;
-                $gScholarPath = $childNode->getAttribute('data-href');
+                $publicationPath = $childNode->getAttribute('data-href');
             } elseif ($cssClass === self::GSCHOLAR_CSS_CLASS_GRAY && $childNodeIndex === 1) {
                 $authors = $childNode->textContent;
             } elseif ($cssClass === self::GSCHOLAR_CSS_CLASS_GRAY && $childNodeIndex === 2) {
@@ -91,7 +91,7 @@ class PublicationParser extends BaseParser implements Parser
 
         return [
             'title' => $title,
-            'gScholarPath' => $gScholarPath,
+            'publicationPath' => $publicationPath,
             'authors' => $authors,
             'publisherDetails' => $publisherDetails,
         ];