+++ CHANGELOG.md
@@ -7,6 +7,16 @@ The format is based on [Keep a Changelog][1], and this project adheres to [Seman
 ### Added
 - Parsing of a scholar profile's statistics.
 
+## [1.0.2] - 2019-03-08
+### Added
+- Publication entity's getter method getPublicationURL().
+### Changed
+- Publication entity's property gScholarPath renamed into publicationPath.
+
+## [1.0.1] - 2019-03-05
+### Added
+- Publication entity's getter methods for all its private properties.
+
 ## [1.0.0] - 2019-03-04
 ### Added
 - Parsing of a scholar profile's publications.